backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
# Atualizado para evitar incompatibilidade gotrue/httpx (erro do "proxy")
supabase==2.28.2
python-dotenv==1.0.1
